Shifted Allocation to client component.
